backend/requirements.txt
fastapi==0.124.1
uvicorn[standard]==0.40.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
websockets==12.0
httpx==0.28.1
bandit==1.8.6
flake8==7.1.0
python-dotenv==1.0.1
PyYAML==6.0.3
